+++ src/components/week/calendarWeekView.component.ts
@@ -16,10 +16,14 @@ import { Subscription } from 'rxjs/Subscription';
 import {
   WeekDay,
   CalendarEvent,
+  WeekViewEvent,
   WeekViewEventRow,
   getWeekViewHeader,
   getWeekView
 } from 'calendar-utils';
+import { ResizeEvent } from 'angular-resizable-element';
+import addDays from 'date-fns/add_days';
+import { CalendarEventTimesChangedEvent } from './../../interfaces/calendarEventTimesChangedEvent.interface';
 
 @Component({
   selector: 'mwl-calendar-week-view',
@@ -34,12 +38,18 @@ import {
           (click)="dayClicked.emit({date: day.date})">
         </mwl-calendar-week-view-header>
       </div>
-      <div *ngFor="let eventRow of eventRows">
+      <div *ngFor="let eventRow of eventRows" #container>
         <div
           class="cal-event-container"
           *ngFor="let weekEvent of eventRow.row"
           [style.width]="((100 / 7) * weekEvent.span) + '%'"
-          [style.marginLeft]="((100 / 7) * weekEvent.offset) + '%'">
+          [style.marginLeft]="((100 / 7) * weekEvent.offset) + '%'"
+          mwlResizable
+          [resizeEdges]="{left: weekEvent.event?.resizable?.beforeStart, right: weekEvent.event?.resizable?.afterEnd}"
+          [resizeSnapGrid]="{left: container.offsetWidth / 7, right: container.offsetWidth / 7}"
+          (resizeStart)="resizeStarted(weekEvent, $event)"
+          (resizing)="resizing(weekEvent, $event, container.offsetWidth / 7)"
+          (resizeEnd)="resizeEnded(weekEvent)">
           <mwl-calendar-week-view-event
             [weekEvent]="weekEvent"
             [tooltipPlacement]="tooltipPlacement"
@@ -92,9 +102,22 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
    */
   @Output() eventClicked: EventEmitter<{event: CalendarEvent}> = new EventEmitter<{event: CalendarEvent}>();
 
+  /**
+   * Called when an event is resized or dragged and dropped
+   */
+  @Output() eventTimesChanged: EventEmitter<CalendarEventTimesChangedEvent> = new EventEmitter<CalendarEventTimesChangedEvent>();
+
   days: WeekDay[];
+
   eventRows: WeekViewEventRow[] = [];
+
   refreshSubscription: Subscription;
+
+  currentResize: {
+    originalOffset: number,
+    originalSpan: number,
+    edge: string
+  };
 
   constructor(private cdr: ChangeDetectorRef, @Inject(LOCALE_ID) locale: string) {
     this.locale = locale;
@@ -125,6 +148,50 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
     if (this.refreshSubscription) {
       this.refreshSubscription.unsubscribe();
     }
+  }
+
+  resizeStarted(weekEvent: WeekViewEvent, resizeEvent: ResizeEvent): void {
+    this.currentResize = {
+      originalOffset: weekEvent.offset,
+      originalSpan: weekEvent.span,
+      edge: typeof resizeEvent.edges.left !== 'undefined' ? 'left' : 'right'
+    };
+  }
+
+  resizing(weekEvent: WeekViewEvent, resizeEvent: ResizeEvent, dayWidth: number): void {
+    if (resizeEvent.edges.left) {
+      const diff: number = Math.round(+resizeEvent.edges.left / dayWidth);
+      weekEvent.offset = this.currentResize.originalOffset + diff;
+      weekEvent.span = this.currentResize.originalSpan - diff;
+    } else if (resizeEvent.edges.right) {
+      const diff: number = Math.round(+resizeEvent.edges.right / dayWidth);
+      weekEvent.span = this.currentResize.originalSpan + diff;
+    }
+
+  }
+
+  resizeEnded(weekEvent: WeekViewEvent): void {
+
+    let daysDiff: number;
+    if (this.currentResize.edge === 'left') {
+      daysDiff = weekEvent.offset - this.currentResize.originalOffset;
+    } else {
+      daysDiff = weekEvent.span - this.currentResize.originalSpan;
+    }
+
+    weekEvent.offset = this.currentResize.originalOffset;
+    weekEvent.span = this.currentResize.originalSpan;
+
+    let newStart: Date = weekEvent.event.start;
+    let newEnd: Date = weekEvent.event.end;
+    if (this.currentResize.edge === 'left') {
+      newStart = addDays(newStart, daysDiff);
+    } else if (newEnd) {
+      newEnd = addDays(newEnd, daysDiff);
+    }
+
+    this.eventTimesChanged.emit({newStart, newEnd, event: weekEvent.event});
+
   }
 
   private refreshHeader(): void {