+++ src/index.umd.ts
@@ -1,2 +1,2 @@
-import './../scss/angular2-calendar.scss';
+import './../scss/angular-calendar.scss';
 export * from './index';